refactor: move catalog and rategrid to 'seismcity' module
